2023-05-09 DB: updates from class
