+++ src/Core/Rooms/Room.php
@@ -1,9 +1,10 @@
 <?php
 
 namespace Core\Rooms;
-use Core\Rooms\AppartmentBlock as AppartmentBlock;
+use Core\Rooms\AppartmentBlock;
 
-class Room extends AppartmentBlock{
+class Room extends AppartmentBlock
+{
 	public function __construct(
 		protected int $number, 
 		protected int $capacity, 
@@ -30,29 +31,4 @@ class Room extends AppartmentBlock{
 		return $this->capacity;
 	}
 }
-
-abstract class AppartmentBlock{
-
-	public function __construct(
-		protected int $number, 
-		protected int $capacity, 
-		protected string $description, 
-		protected float $price) {}
-    	
-	abstract public function getNumber() : int;
-	
-	abstract public function getCapacity() : int;
-	
-	abstract protected function createRoom(): Room;
-}
-
-class Appartment extends AppartmentBlock {
-    protected function createRoom(): Room {
-        return new Room(1, 2, "A cozy bedroom", 1000.00);
-    }
-}
-
-$room = new AppartmentBlock(101, 2, "closet", 1200);
-var_dump($room->createRoom());
-
